class XML_Directory : - added getFiles(), flat browsing with extensions selector - added addDirectory(), add a sub-directory in the the current directory class XML_File : - added getLastChange(), date of the last modification - extended xml parsing by adding rights attributes
new function booltostr, transform a boolean in the equivalent string (true or false)
+ various debug...

+++ trunk/lib/XML.php
@@ -277,7 +277,7 @@ class XML_Document extends DOMDocument {
       
       $sPath = MAIN_DIRECTORY.$sPath;
       
-      unlink($sPath);
+      if (file_exists($sPath)) unlink($sPath);
       parent::save($sPath);
       
       return true;
@@ -835,6 +835,7 @@ class XML_Element extends DOMElement {
   public function setAttribute($sName, $sValue = '') {
     
     // TODO : RIGHTS
+    
     if ($sValue !== '' && $sValue !== null) return parent::setAttribute($sName, $sValue);
     else return $this->removeAttribute($sName);
   }
